Rename mongo_data.go to mongo_datum.go and change DataRepository to DatumRepository for separate DataSet and Datum repositories.
Move deviceData with type='upload' to its own collection named deviceDataSets.

Collection deviceDataSets acts as a collection for "parent" datum where its
children are in the original deviceData collection. Split
platform/data/store.DataRepository into 2 repositories, one for DataSet and one
for data of these DataSets called DataSetRepository and DatumRepository,
respectively. Duplicate writes in a transaction for creating/modifying DataSets
so that any clients/services still using the old collection will work properly.
Read from both old, deviceData and new, deviceDataSets collections when reading
a dataset and return the first that contains something.

Remove dual writes for methods that don't need it.
