+++ data/store/mongo/mongo_test.go
@@ -264,6 +264,7 @@ var _ = Describe("Mongo", func() {
 
 	Context("with a new store", func() {
 		var collection *mongo.Collection
+		var dataSetCollection *mongo.Collection
 		var summaryCollection *mongo.Collection
 
 		BeforeEach(func() {
@@ -272,14 +273,16 @@ var _ = Describe("Mongo", func() {
 			Expect(err).ToNot(HaveOccurred())
 			Expect(store).ToNot(BeNil())
 			collection = store.GetCollection("deviceData")
+			dataSetCollection = store.GetCollection("deviceDataSets")
 			summaryCollection = store.GetCollection("summary")
 			Expect(store.EnsureIndexes()).To(Succeed())
 		})
 
 		AfterEach(func() {
 			if collection != nil {
-				_ = collection.Database().Drop(context.Background())
-				_ = summaryCollection.Database().Drop(context.Background())
+				collection.Database().Drop(context.Background())
+				dataSetCollection.Database().Drop(context.Background())
+				summaryCollection.Database().Drop(context.Background())
 			}
 		})
 
@@ -406,22 +409,23 @@ var _ = Describe("Mongo", func() {
 				preparePersistedDataSets := func() {
 					createdTimeOther, _ := time.Parse(time.RFC3339, "2016-09-01T12:00:00Z")
 					collection = store.GetCollection("deviceData")
+					dataSetCollection = store.GetCollection("deviceDataSets")
 					dataSetExistingOther = NewDataSet(userTest.RandomID(), dataTest.NewDeviceID())
 					dataSetExistingOther.CreatedTime = pointer.FromTime(createdTimeOther)
 					dataSetExistingOther.ModifiedTime = pointer.FromTime(createdTimeOther)
-					_, err := collection.InsertOne(context.Background(), dataSetExistingOther)
+					_, err := dataSetCollection.InsertOne(context.Background(), dataSetExistingOther)
 					Expect(err).ToNot(HaveOccurred())
 					dataSetExistingOne = NewDataSet(userID, deviceID)
 					createdTimeOne, _ := time.Parse(time.RFC3339, "2016-09-01T12:30:00Z")
 					dataSetExistingOne.CreatedTime = pointer.FromTime(createdTimeOne)
 					dataSetExistingOne.ModifiedTime = pointer.FromTime(createdTimeOne)
-					_, err = collection.InsertOne(context.Background(), dataSetExistingOne)
+					_, err = dataSetCollection.InsertOne(context.Background(), dataSetExistingOne)
 					Expect(err).ToNot(HaveOccurred())
 					dataSetExistingTwo = NewDataSet(userID, deviceID)
 					createdTimeTwo, _ := time.Parse(time.RFC3339, "2016-09-01T10:00:00Z")
 					dataSetExistingTwo.CreatedTime = pointer.FromTime(createdTimeTwo)
 					dataSetExistingTwo.ModifiedTime = pointer.FromTime(createdTimeTwo)
-					_, err = collection.InsertOne(context.Background(), dataSetExistingTwo)
+					_, err = dataSetCollection.InsertOne(context.Background(), dataSetExistingTwo)
 					Expect(err).ToNot(HaveOccurred())
 				}
 
@@ -473,7 +477,7 @@ var _ = Describe("Mongo", func() {
 					Context("with database access", func() {
 						BeforeEach(func() {
 							preparePersistedDataSets()
-							_, err := collection.InsertOne(context.Background(), dataSet)
+							_, err := dataSetCollection.InsertOne(context.Background(), dataSet)
 							Expect(err).ToNot(HaveOccurred())
 						})
 
@@ -511,7 +515,7 @@ var _ = Describe("Mongo", func() {
 							BeforeEach(func() {
 								createdTime, _ := time.Parse(time.RFC3339, "2016-09-01T13:00:00Z")
 								dataSet.DeletedTime = pointer.FromTime(createdTime)
-								result := collection.FindOneAndReplace(context.Background(), bson.M{"id": dataSet.ID}, dataSet)
+								result := dataSetCollection.FindOneAndReplace(context.Background(), bson.M{"id": dataSet.ID}, dataSet)
 								Expect(result.Err()).ToNot(HaveOccurred())
 							})
 
@@ -544,7 +548,7 @@ var _ = Describe("Mongo", func() {
 					Context("with database access", func() {
 						BeforeEach(func() {
 							preparePersistedDataSets()
-							_, err := collection.InsertOne(context.Background(), dataSet)
+							_, err := dataSetCollection.InsertOne(context.Background(), dataSet)
 							Expect(err).ToNot(HaveOccurred())
 						})
 
@@ -616,7 +620,7 @@ var _ = Describe("Mongo", func() {
 
 							// Make sure the values are set in the db as well.
 							var result *upload.Upload
-							err := collection.FindOne(context.Background(), bson.M{"uploadId": dataSet.UploadID}).Decode(&result)
+							err := dataSetCollection.FindOne(context.Background(), bson.M{"uploadId": dataSet.UploadID}).Decode(&result)
 							Expect(err).ToNot(HaveOccurred())
 							Expect(*result.CreatedTime).To(Equal(*dataSet.CreatedTime))
 							Expect(*result.ModifiedTime).To(Equal(*dataSet.ModifiedTime))
@@ -625,9 +629,9 @@ var _ = Describe("Mongo", func() {
 						})
 
 						It("has the correct stored data sets", func() {
-							ValidateDataSet(collection, bson.M{"createdTime": bson.M{"$exists": true}, "createdUserId": bson.M{"$exists": false}}, bson.M{}, dataSetExistingOther, dataSetExistingOne, dataSetExistingTwo)
+							ValidateDataSet(dataSetCollection, bson.M{"createdTime": bson.M{"$exists": true}, "createdUserId": bson.M{"$exists": false}}, bson.M{}, dataSetExistingOther, dataSetExistingOne, dataSetExistingTwo)
 							Expect(repository.CreateDataSet(ctx, dataSet)).To(Succeed())
-							ValidateDataSet(collection, bson.M{"createdTime": bson.M{"$exists": true}, "createdUserId": bson.M{"$exists": false}}, bson.M{}, dataSetExistingOther, dataSetExistingOne, dataSetExistingTwo, dataSet)
+							ValidateDataSet(dataSetCollection, bson.M{"createdTime": bson.M{"$exists": true}, "createdUserId": bson.M{"$exists": false}}, bson.M{}, dataSetExistingOther, dataSetExistingOne, dataSetExistingTwo, dataSet)
 						})
 					})
 				})
@@ -681,13 +685,20 @@ var _ = Describe("Mongo", func() {
 							createdTime := time.Now().UTC().Truncate(time.Millisecond)
 							dataSet.CreatedTime = pointer.FromTime(createdTime)
 							dataSet.ModifiedTime = pointer.FromTime(createdTime)
-							_, err := collection.InsertOne(context.Background(), dataSet)
+							// Insert in BOTH collections to mimick the
+							// migration where dataSet will be in deviceData
+							// and deviceDataSets. This is because while
+							// migration happens an update to a dataset will
+							// only succeed if it is still in the old deviceData collection.
+							_, err := dataSetCollection.InsertOne(context.Background(), dataSet)
+							Expect(err).ToNot(HaveOccurred())
+							_, err = collection.InsertOne(context.Background(), dataSet)
 							Expect(err).ToNot(HaveOccurred())
 							id = *dataSet.UploadID
 						})
 
 						AfterEach(func() {
-							logger.AssertDebug("UpdateDataSet", log.Fields{"id": id, "update": update})
+							logger.AssertDebug("DatumRepository.UpdateDataSet", log.Fields{"id": id, "update": update})
 						})
 
 						Context("with updates", func() {
@@ -724,9 +735,9 @@ var _ = Describe("Mongo", func() {
 						})
 
 						It("has the correct stored data sets", func() {
-							ValidateDataSet(collection, bson.M{}, bson.M{}, dataSetExistingOther, dataSetExistingOne, dataSetExistingTwo, dataSet)
+							ValidateDataSet(dataSetCollection, bson.M{}, bson.M{}, dataSetExistingOther, dataSetExistingOne, dataSetExistingTwo, dataSet)
 							// All newly created data now includes the modifiedTime as well.
-							ValidateDataSet(collection, bson.M{"modifiedTime": bson.M{"$exists": true}}, bson.M{}, dataSetExistingOther, dataSetExistingOne, dataSetExistingTwo, dataSet)
+							ValidateDataSet(dataSetCollection, bson.M{"modifiedTime": bson.M{"$exists": true}}, bson.M{}, dataSetExistingOther, dataSetExistingOne, dataSetExistingTwo, dataSet)
 							update = data.NewDataSetUpdate()
 							update.State = pointer.FromString("closed")
 							result, err := repository.UpdateDataSet(ctx, id, update)
@@ -734,8 +745,15 @@ var _ = Describe("Mongo", func() {
 							Expect(result).ToNot(BeNil())
 							Expect(result.State).ToNot(BeNil())
 							Expect(*result.State).To(Equal("closed"))
-							ValidateDataSet(collection, bson.M{}, bson.M{}, dataSetExistingOther, dataSetExistingOne, dataSetExistingTwo, result)
-							ValidateDataSet(collection, bson.M{"modifiedTime": bson.M{"$exists": true}}, bson.M{}, dataSetExistingOther, dataSetExistingOne, dataSetExistingTwo, result)
+							ValidateDataSet(dataSetCollection, bson.M{}, bson.M{}, dataSetExistingOther, dataSetExistingOne, dataSetExistingTwo, result)
+							ValidateDataSet(dataSetCollection, bson.M{"modifiedTime": bson.M{"$exists": true}}, bson.M{}, dataSetExistingOther, dataSetExistingOne, dataSetExistingTwo, result)
+
+							// Updating an existing dataset that lives in the
+							// previous deviceData collection via
+							// DatumRepository should also call
+							// DataSetRepository.upsertDataSet to upsert the
+							// data set into the new deviceDataSets repo
+							logger.AssertDebug("DataSetRepository.upsertDataSet", log.Fields{"id": id, "update": update})
 						})
 					})
 				})
@@ -748,7 +766,14 @@ var _ = Describe("Mongo", func() {
 
 					preparePersistedDataSetsData := func() {
 						preparePersistedDataSets()
+						// Insert in BOTH collections to mimick the
+						// migration where dataSet will be in deviceData
+						// and deviceDataSets. This is because while
+						// migration happens an update to a dataset will
+						// only succeed if it is still in the old deviceData collection.
 						_, err := collection.InsertOne(context.Background(), dataSet)
+						Expect(err).ToNot(HaveOccurred())
+						_, err = dataSetCollection.InsertOne(context.Background(), dataSet)
 						Expect(err).ToNot(HaveOccurred())
 						Expect(repository.CreateDataSetData(ctx, dataSetExistingOther, dataSetExistingOtherData)).To(Succeed())
 						Expect(repository.CreateDataSetData(ctx, dataSetExistingOne, dataSetExistingOneData)).To(Succeed())
@@ -809,9 +834,9 @@ var _ = Describe("Mongo", func() {
 							})
 
 							It("has the correct stored data sets", func() {
-								ValidateDataSet(collection, bson.M{"deletedTime": bson.M{"$exists": true}, "deletedUserId": bson.M{"$exists": false}}, bson.M{})
+								ValidateDataSet(dataSetCollection, bson.M{"deletedTime": bson.M{"$exists": true}, "deletedUserId": bson.M{"$exists": false}}, bson.M{})
 								Expect(repository.DeleteDataSet(ctx, dataSet)).To(Succeed())
-								ValidateDataSet(collection, bson.M{"deletedTime": bson.M{"$exists": true}, "deletedUserId": bson.M{"$exists": false}}, bson.M{}, dataSet)
+								ValidateDataSet(dataSetCollection, bson.M{"deletedTime": bson.M{"$exists": true}, "deletedUserId": bson.M{"$exists": false}}, bson.M{}, dataSet)
 							})
 
 							It("has the correct stored data set data", func() {
@@ -969,14 +994,14 @@ var _ = Describe("Mongo", func() {
 										preparePersistedDataSetsData()
 										Expect(repository.CreateDataSetData(ctx, dataSet, dataSetData)).To(Succeed())
 										ValidateDataSetData(collection, bson.M{"_active": true}, bson.M{}, data.Data{})
-										ValidateDataSet(collection, bson.M{"uploadId": dataSet.UploadID}, bson.M{}, dataSet)
+										ValidateDataSet(dataSetCollection, bson.M{"uploadId": dataSet.UploadID}, bson.M{}, dataSet)
 									})
 
 									It("succeeds and has the correct stored active data set data", func() {
 										Expect(repository.ActivateDataSetData(ctx, dataSet, selectors)).To(Succeed())
 										selectedDataSetData.SetActive(true)
 										ValidateDataSetData(collection, bson.M{"_active": true}, bson.M{"modifiedTime": 0}, selectedDataSetData)
-										ValidateDataSet(collection, bson.M{"uploadId": dataSet.UploadID}, bson.M{}, dataSet)
+										ValidateDataSet(dataSetCollection, bson.M{"uploadId": dataSet.UploadID}, bson.M{}, dataSet)
 									})
 
 									It("succeeds with no changes when the data set user id is different", func() {
@@ -1106,14 +1131,14 @@ var _ = Describe("Mongo", func() {
 										Expect(repository.ActivateDataSetData(ctx, dataSet, nil)).To(Succeed())
 										dataSetData.SetActive(true)
 										ValidateDataSetData(collection, bson.M{"_active": false, "uploadId": dataSet.UploadID}, bson.M{"modifiedTime": 0}, data.Data{})
-										ValidateDataSet(collection, bson.M{"uploadId": dataSet.UploadID}, bson.M{}, dataSet)
+										ValidateDataSet(dataSetCollection, bson.M{"uploadId": dataSet.UploadID}, bson.M{}, dataSet)
 									})
 
 									It("succeeds and has the correct stored active data set data", func() {
 										Expect(repository.ArchiveDataSetData(ctx, dataSet, selectors)).To(Succeed())
 										selectedDataSetData.SetActive(false)
 										ValidateDataSetData(collection, bson.M{"_active": false, "uploadId": dataSet.UploadID}, bson.M{"archivedTime": 0, "modifiedTime": 0}, selectedDataSetData)
-										ValidateDataSet(collection, bson.M{"uploadId": dataSet.UploadID}, bson.M{}, dataSet)
+										ValidateDataSet(dataSetCollection, bson.M{"uploadId": dataSet.UploadID}, bson.M{}, dataSet)
 									})
 
 									It("succeeds and updates .modifiedTime", func() {
@@ -1250,14 +1275,14 @@ var _ = Describe("Mongo", func() {
 										preparePersistedDataSetsData()
 										Expect(repository.CreateDataSetData(ctx, dataSet, dataSetData)).To(Succeed())
 										ValidateDataSetData(collection, bson.M{"deletedTime": bson.M{"$exists": true}}, bson.M{"modifiedTime": 0}, data.Data{})
-										ValidateDataSet(collection, bson.M{"uploadId": dataSet.UploadID}, bson.M{}, dataSet)
+										ValidateDataSet(dataSetCollection, bson.M{"uploadId": dataSet.UploadID}, bson.M{}, dataSet)
 									})
 
 									It("succeeds and has the correct stored active data set data", func() {
 										Expect(repository.DeleteDataSetData(ctx, dataSet, selectors)).To(Succeed())
 										selectedDataSetData.SetActive(false)
 										ValidateDataSetData(collection, bson.M{"deletedTime": bson.M{"$exists": true}}, bson.M{"archivedTime": 0, "deletedTime": 0, "modifiedTime": 0}, selectedDataSetData)
-										ValidateDataSet(collection, bson.M{"uploadId": dataSet.UploadID}, bson.M{}, dataSet)
+										ValidateDataSet(dataSetCollection, bson.M{"uploadId": dataSet.UploadID}, bson.M{}, dataSet)
 									})
 
 									It("succeeds with no changes when the data set user id is different", func() {
@@ -1386,13 +1411,13 @@ var _ = Describe("Mongo", func() {
 										Expect(repository.CreateDataSetData(ctx, dataSet, dataSetData)).To(Succeed())
 										Expect(repository.DeleteDataSetData(ctx, dataSet, nil)).To(Succeed())
 										ValidateDataSetData(collection, bson.M{"deletedTime": bson.M{"$exists": true}}, bson.M{"archivedTime": 0, "deletedTime": 0, "modifiedTime": 0}, dataSetData)
-										ValidateDataSet(collection, bson.M{"uploadId": dataSet.UploadID}, bson.M{}, dataSet)
+										ValidateDataSet(dataSetCollection, bson.M{"uploadId": dataSet.UploadID}, bson.M{}, dataSet)
 									})
 
 									It("succeeds and has the correct stored active data set data", func() {
 										Expect(repository.DestroyDeletedDataSetData(ctx, dataSet, selectors)).To(Succeed())
 										ValidateDataSetData(collection, bson.M{"deletedTime": bson.M{"$exists": true}}, bson.M{"archivedTime": 0, "deletedTime": 0, "modifiedTime": 0}, unselectedDataSetData)
-										ValidateDataSet(collection, bson.M{"uploadId": dataSet.UploadID}, bson.M{}, dataSet)
+										ValidateDataSet(dataSetCollection, bson.M{"uploadId": dataSet.UploadID}, bson.M{}, dataSet)
 									})
 
 									It("succeeds with no changes when the data set user id is different", func() {
@@ -1520,13 +1545,13 @@ var _ = Describe("Mongo", func() {
 										preparePersistedDataSetsData()
 										Expect(repository.CreateDataSetData(ctx, dataSet, dataSetData)).To(Succeed())
 										ValidateDataSetData(collection, bson.M{"uploadId": dataSet.UploadID}, bson.M{"modifiedTime": 0}, dataSetData)
-										ValidateDataSet(collection, bson.M{"uploadId": dataSet.UploadID}, bson.M{}, dataSet)
+										ValidateDataSet(dataSetCollection, bson.M{"uploadId": dataSet.UploadID}, bson.M{}, dataSet)
 									})
 
 									It("succeeds and has the correct stored active data set data", func() {
 										Expect(repository.DestroyDataSetData(ctx, dataSet, selectors)).To(Succeed())
 										ValidateDataSetData(collection, bson.M{"uploadId": dataSet.UploadID}, bson.M{"modifiedTime": 0}, unselectedDataSetData)
-										ValidateDataSet(collection, bson.M{"uploadId": dataSet.UploadID}, bson.M{}, dataSet)
+										ValidateDataSet(dataSetCollection, bson.M{"uploadId": dataSet.UploadID}, bson.M{}, dataSet)
 									})
 
 									It("succeeds with no changes when the data set user id is different", func() {
@@ -1673,9 +1698,9 @@ var _ = Describe("Mongo", func() {
 							})
 
 							It("has the correct stored data sets", func() {
-								ValidateDataSet(collection, bson.M{"uploadId": dataSetExistingOne.UploadID, "_active": true}, bson.M{}, dataSetExistingOne)
+								ValidateDataSet(dataSetCollection, bson.M{"uploadId": dataSetExistingOne.UploadID, "_active": true}, bson.M{}, dataSetExistingOne)
 								Expect(repository.ArchiveDeviceDataUsingHashesFromDataSet(ctx, dataSet)).To(Succeed())
-								ValidateDataSet(collection, bson.M{"uploadId": dataSetExistingOne.UploadID, "_active": true}, bson.M{}, dataSetExistingOne)
+								ValidateDataSet(dataSetCollection, bson.M{"uploadId": dataSetExistingOne.UploadID, "_active": true}, bson.M{}, dataSetExistingOne)
 							})
 
 							It("has the correct stored archived data set data", func() {
@@ -1759,9 +1784,9 @@ var _ = Describe("Mongo", func() {
 							})
 
 							It("has the correct stored data sets", func() {
-								ValidateDataSet(collection, bson.M{"uploadId": dataSetExistingOne.UploadID, "_active": true}, bson.M{}, dataSetExistingOne)
+								ValidateDataSet(dataSetCollection, bson.M{"uploadId": dataSetExistingOne.UploadID, "_active": true}, bson.M{}, dataSetExistingOne)
 								Expect(repository.UnarchiveDeviceDataUsingHashesFromDataSet(ctx, dataSet)).To(Succeed())
-								ValidateDataSet(collection, bson.M{"uploadId": dataSetExistingOne.UploadID, "_active": true}, bson.M{}, dataSetExistingOne)
+								ValidateDataSet(dataSetCollection, bson.M{"uploadId": dataSetExistingOne.UploadID, "_active": true}, bson.M{}, dataSetExistingOne)
 							})
 
 							It("has the correct stored unarchived data set data", func() {
@@ -1786,9 +1811,9 @@ var _ = Describe("Mongo", func() {
 							})
 
 							It("has the correct stored data sets if an intermediary is unarchived", func() {
-								ValidateDataSet(collection, bson.M{"uploadId": dataSetExistingTwo.UploadID, "_active": true}, bson.M{}, dataSetExistingTwo)
+								ValidateDataSet(dataSetCollection, bson.M{"uploadId": dataSetExistingTwo.UploadID, "_active": true}, bson.M{}, dataSetExistingTwo)
 								Expect(repository.UnarchiveDeviceDataUsingHashesFromDataSet(ctx, dataSetExistingOne)).To(Succeed())
-								ValidateDataSet(collection, bson.M{"uploadId": dataSetExistingTwo.UploadID, "_active": true}, bson.M{}, dataSetExistingTwo)
+								ValidateDataSet(dataSetCollection, bson.M{"uploadId": dataSetExistingTwo.UploadID, "_active": true}, bson.M{}, dataSetExistingTwo)
 							})
 
 							It("has the correct stored unarchived data set data if an intermediary is unarchived", func() {
@@ -1865,12 +1890,12 @@ var _ = Describe("Mongo", func() {
 							})
 
 							It("has the correct stored active data set", func() {
-								ValidateDataSet(collection, bson.M{}, bson.M{}, dataSet, dataSetExistingOther, dataSetExistingOne, dataSetExistingTwo)
-								ValidateDataSet(collection, bson.M{"deletedTime": bson.M{"$exists": false}, "deletedUserId": bson.M{"$exists": false}}, bson.M{}, dataSet, dataSetExistingOther, dataSetExistingOne, dataSetExistingTwo)
+								ValidateDataSet(dataSetCollection, bson.M{}, bson.M{}, dataSet, dataSetExistingOther, dataSetExistingOne, dataSetExistingTwo)
+								ValidateDataSet(dataSetCollection, bson.M{"deletedTime": bson.M{"$exists": false}, "deletedUserId": bson.M{"$exists": false}}, bson.M{}, dataSet, dataSetExistingOther, dataSetExistingOne, dataSetExistingTwo)
 								Expect(repository.DeleteOtherDataSetData(ctx, dataSet)).To(Succeed())
-								Expect(collection.CountDocuments(ctx, bson.M{"type": "upload"})).To(Equal(int64(4)))
-								ValidateDataSet(collection, bson.M{"deletedTime": bson.M{"$exists": true}, "deletedUserId": bson.M{"$exists": false}}, bson.M{"deletedTime": 0}, dataSetExistingTwo, dataSetExistingOne)
-								ValidateDataSet(collection, bson.M{"deletedTime": bson.M{"$exists": false}, "deletedUserId": bson.M{"$exists": false}}, bson.M{}, dataSet, dataSetExistingOther)
+								Expect(dataSetCollection.CountDocuments(ctx, bson.M{"type": "upload"})).To(Equal(int64(4)))
+								ValidateDataSet(dataSetCollection, bson.M{"deletedTime": bson.M{"$exists": true}, "deletedUserId": bson.M{"$exists": false}}, bson.M{"deletedTime": 0}, dataSetExistingTwo, dataSetExistingOne)
+								ValidateDataSet(dataSetCollection, bson.M{"deletedTime": bson.M{"$exists": false}, "deletedUserId": bson.M{"$exists": false}}, bson.M{}, dataSet, dataSetExistingOther)
 							})
 
 							It("has the correct stored active data set data", func() {
@@ -1907,7 +1932,14 @@ var _ = Describe("Mongo", func() {
 								createdTime, _ := time.Parse(time.RFC3339, "2016-09-01T11:00:00Z")
 								destroyDataSet.CreatedTime = pointer.FromTime(createdTime)
 								destroyDataSet.ModifiedTime = pointer.FromTime(createdTime)
+								// Insert in BOTH collections to mimick the
+								// migration where dataSet will be in deviceData
+								// and deviceDataSets. This is because while
+								// migration happens an update to a dataset will
+								// only succeed if it is still in the old deviceData collection.
 								_, err := collection.InsertOne(context.Background(), destroyDataSet)
+								Expect(err).ToNot(HaveOccurred())
+								_, err = dataSetCollection.InsertOne(context.Background(), destroyDataSet)
 								Expect(err).ToNot(HaveOccurred())
 								destroyDataSetData = NewDataSetData(destroyDeviceID)
 								Expect(repository.CreateDataSetData(ctx, destroyDataSet, destroyDataSetData)).To(Succeed())
@@ -1918,9 +1950,9 @@ var _ = Describe("Mongo", func() {
 							})
 
 							It("has the correct stored data sets", func() {
-								ValidateDataSet(collection, bson.M{}, bson.M{}, dataSet, dataSetExistingOther, dataSetExistingOne, dataSetExistingTwo, destroyDataSet)
+								ValidateDataSet(dataSetCollection, bson.M{}, bson.M{}, dataSet, dataSetExistingOther, dataSetExistingOne, dataSetExistingTwo, destroyDataSet)
 								Expect(repository.DestroyDataForUserByID(ctx, destroyUserID)).To(Succeed())
-								ValidateDataSet(collection, bson.M{}, bson.M{}, dataSet, dataSetExistingOther, dataSetExistingOne, dataSetExistingTwo)
+								ValidateDataSet(dataSetCollection, bson.M{}, bson.M{}, dataSet, dataSetExistingOther, dataSetExistingOne, dataSetExistingTwo)
 							})
 
 							It("has the correct stored data set data", func() {
